+++ src/KMonad/Args/Parser.hs
@@ -28,6 +28,7 @@ module KMonad.Args.Parser
   , otokens
   , itokens
   , buttonP
+  , implArndButtons
   , keywordButtons
   , noKeywordButtons
   )
@@ -183,7 +184,7 @@ exprP = paren . choice $
 
 -- | Different ways to refer to shifted versions of keycodes
 shiftedNames :: [(Text, DefButton)]
-shiftedNames = let f = second $ \kc -> KAround (KEmit KeyLeftShift) (KEmit kc) in
+shiftedNames = let f = second $ \kc -> KAroundImplicit (KEmit KeyLeftShift) (KEmit kc) in
                  map f $ cps <> num <> oth <> lng
   where
     cps = zip (map T.singleton ['A'..'Z'])
@@ -201,7 +202,7 @@ shiftedNames = let f = second $ \kc -> KAround (KEmit KeyLeftShift) (KEmit kc) i
 buttonNames :: [(Text, DefButton)]
 buttonNames = shiftedNames <> escp <> util
   where
-    emitS c = KAround (KEmit KeyLeftShift) (KEmit c)
+    emitS c = KAroundImplicit (KEmit KeyLeftShift) (KEmit c)
     -- Escaped versions for reserved characters
     escp = [ ("\\(", emitS Key9), ("\\)", emitS Key0)
            , ("\\_", emitS KeyMinus), ("\\\\", KEmit KeyBackslash)]
@@ -213,7 +214,7 @@ buttonNames = shiftedNames <> escp <> util
 
 -- | Parse "X-b" style modded-sequences
 moddedP :: Parser DefButton
-moddedP = KAround <$> prfx <*> buttonP
+moddedP = KAroundImplicit <$> prfx <*> buttonP
   where mods = [ ("S-", KeyLeftShift), ("C-", KeyLeftCtrl)
                , ("A-", KeyLeftAlt),   ("M-", KeyLeftMeta)
                , ("RS-", KeyRightShift), ("RC-", KeyRightCtrl)
@@ -264,7 +265,7 @@ buttonP = (lexeme . choice . map try $
 -- @(keyword, how to parse the token)@
 keywordButtons :: [(Text, Parser DefButton)]
 keywordButtons =
-  [ ("around"         , KAround      <$> buttonP     <*> buttonP)
+  [ ("around-implicit", KAroundImplicit <$> buttonP <*> buttonP)
   , ("press-only"     , KPressOnly   <$> keycodeP)
   , ("release-only"   , KReleaseOnly <$> keycodeP)
   , ("multi-tap"      , KMultiTap    <$> timed       <*> buttonP)
@@ -301,9 +302,17 @@ keywordButtons =
   , ("pause"          , KPause . fromIntegral <$> numP)
   , ("sticky-key"     , KStickyKey   <$> lexeme numP <*> buttonP)
   ]
+  ++ map (\(nm,_,btn) -> (nm, btn <$> buttonP <*> buttonP)) implArndButtons
  where
   timed :: Parser [(Int, DefButton)]
   timed = many ((,) <$> lexeme numP <*> lexeme buttonP)
+
+implArndButtons :: [(Text, ImplArnd, DefButton -> DefButton -> DefButton)]
+implArndButtons =
+  [ ("around"           , IAAround         , KAround)
+  , ("around-only"      , IAAroundOnly     , KAroundOnly)
+  , ("around-when-alone", IAAroundWhenAlone, KAroundWhenAlone)
+  ]
 
 -- | Parsers for buttons that do __not__ have a keyword at the start
 noKeywordButtons :: [Parser DefButton]
@@ -343,6 +352,12 @@ otokens =
   , ("send-event-sink", KSendEventSink <$> optional ((,) <$> lexeme numP <*> lexeme numP))
   , ("kext"           , pure KKextSink)]
 
+-- | Parse an impl arnd token
+implArndP :: Parser ImplArnd
+implArndP = lexeme . choice $
+  try (IADisabled <$ symbol "disabled")
+  : map (\(s, v, _) -> try $ v <$ symbol s) implArndButtons
+
 -- | Parse the DefCfg token
 defcfgP :: Parser DefSettings
 defcfgP = some (lexeme settingP)
@@ -357,6 +372,7 @@ settingP = let f s p = symbol s *> p in
     , SFallThrough <$> f "fallthrough"   bool
     , SAllowCmd    <$> f "allow-cmd"     bool
     , SCmpSeqDelay <$> f "cmp-seq-delay" numP
+    , SImplArnd    <$> f "implicit-around" implArndP
     ])
 
 --------------------------------------------------------------------------------
@@ -379,8 +395,12 @@ defsrcP =
 -- $deflayer
 
 deflayerP :: Parser DefLayer
-deflayerP =
-  DefLayer
-    <$> lexeme word
-    <*> optional (keywordP "source" word)
-    <*> many (lexeme buttonP)
+deflayerP = DefLayer <$> lexeme word <*> many (lexeme layerSettingP)
+
+layerSettingP :: Parser DefLayerSetting
+layerSettingP =
+  lexeme . choice . map try $
+    [ LSrcName <$> keywordP "source" word
+    , LImplArnd <$> keywordP "implicit-around" implArndP
+    , LButton <$> buttonP
+    ]